rewritting assignment2
