--- Assignment 2(Failed)/src/location/Building.java
@@ -1,7 +1,0 @@
-
-package location;
-public abstract class Building {
-    // Method signature for drawing an object
-    public abstract String getAddress();
-    public  abstract String getContactinfo();
-}
