+++ Assignment 2/src/location/Business.java
@@ -1,12 +1,14 @@
 package location;
 
-public abstract class Building {
+public abstract class Business {
+    /*Description: class that implements basic requirements for any building
+     *
+     * */
 
     public abstract String getContactInfo();
+
     public abstract String getAddress();
 
     public abstract int size();
-
-    public abstract boolean atCapacity();
 
 }
