+++ Assignment 2/src/person/Employee.java
@@ -1,12 +1,18 @@
 package person;
 
 import location.Department;
-import location.Floor;
 import location.Hospital;
 
-public abstract class Employee extends Person implements HospitalLocation{
+public abstract class Employee extends Person implements HospitalLocation {
+    /*Description: stores data of a hired personel at the hosptial. This includes the data of
+     * its parent (Person) while also an assigned id and hospital/department.
+     *
+     * Args: id: unique id given to all staff
+     *       hospital: Hospital of employee
+     *       department: department of employee
+     * */
 
-    private Hospital hospital;
+    private final Hospital hospital;
     private Department department;
 
     private int id;
@@ -19,33 +25,33 @@ public abstract class Employee extends Person implements HospitalLocation{
     }
 
     @Override
-    public String getName(){
+    public String getName() {
         return "(" + this.getId() + ") " + this.getFName() + " " + this.getLname();
     }
 
     @Override
-    public String toString(){
+    public String toString() {
         return this.getName();
     }
 
     @Override
-    public int hashCode(){
+    public int hashCode() {
         return this.getName().hashCode();
     }
 
     @Override
-    public boolean equals(Object obj){
-        if ( (obj == null) || (obj.getClass() != this.getClass()) ){
+    public boolean equals(Object obj) {
+        if ((obj == null) || (obj.getClass() != this.getClass())) {
             return false;
         }
         return obj.hashCode() == this.hashCode();
     }
 
-    public Integer getId(){
+    public Integer getId() {
         return this.id;
     }
 
-    public void setId(Integer id){
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -53,15 +59,15 @@ public abstract class Employee extends Person implements HospitalLocation{
     public Hospital getHospital() {
         return this.hospital;
     }
+
     @Override
-    public Department getDepartment(){
+    public Department getDepartment() {
         return this.department;
     }
 
-    public void setDepartment(Department department){
+    public void setDepartment(Department department) {
         this.department = department;
     }
-
 
 
 }
