+++ Assignment 2/src/person/HospitalLocation.java
@@ -1,0 +1,13 @@
+package person;
+
+import location.Department;
+import location.Hospital;
+
+public interface HospitalLocation {
+    /*description: information required for any hosptial personal.
+     *
+     * */
+    Hospital getHospital();
+
+    Department getDepartment();
+}
