--- Assingment2/src/person/HospitalLocation.java
@@ -1,10 +1,0 @@
-package person;
-
-import location.Department;
-import location.Floor;
-import location.Hospital;
-
-public interface HospitalLocation {
-    public Hospital getHospital();
-    public Department getDepartment();
-}
